feat(Modal): add oversize option and separate out nested components
