+++ src/Modal/Close/index.ts
@@ -1,0 +1,1 @@
+export { ModalClose } from './Close.component';
