+++ src/Modal/Content/index.ts
@@ -1,0 +1,1 @@
+export { ModalContent, StyledContent } from './Content.component';
