+++ src/Modal/Footer/index.ts
@@ -1,0 +1,1 @@
+export { ModalFooter, StyledFooter } from './Footer.component';
