+++ src/Modal/Header/index.ts
@@ -1,0 +1,1 @@
+export { ModalHeader, StyledHeader } from './Header.component';
