+++ src/Modal/Modal.component.tsx
@@ -1,18 +1,21 @@
 // REACT
 import * as React from 'react';
 
-// ANCHOR
-import { Typography } from '..';
-import { Close } from '../Icon';
-import { IconSVGProps } from '../Icon/utils';
-import { sizes } from '../theme/sizes.theme';
-import { colors } from '../theme/colors.theme';
-
 // VENDOR
 import * as StyledReactModal from 'styled-react-modal';
 import styled, { css } from 'styled-components';
 import classnames from 'classnames';
 import { opacify } from 'polished';
+
+// ANCHOR
+import { sizes } from '../theme/sizes.theme';
+import { colors } from '../theme/colors.theme';
+
+// SUBCOMPONENTS
+import { ModalHeader, StyledHeader } from './Header';
+import { ModalContent, StyledContent } from './Content';
+import { ModalFooter, StyledFooter } from './Footer';
+import { ModalClose } from './Close';
 
 export type ModalSize = 'lg' | 'sm';
 
@@ -20,158 +23,6 @@ const Sizes = {
     lg: { width: 50, contentPadding: 2.5, footerHeight: 6.5 },
     sm: { width: 25, contentPadding: 2, footerHeight: 5.5 },
 };
-const borderRadius = sizes.border.radius.modal;
-
-// Modal.Content
-// ------------------------------------------------------------------------------------------------------------------
-
-const StyledContent = styled.div<ModalContentProps>`
-    box-sizing: border-box;
-    width: 100%;
-
-    order: 0;
-    flex-grow: 1;
-`;
-
-interface ModalContentProps {
-    children?: any;
-    className?: string;
-}
-
-const ModalContent = ({ children, className }: ModalContentProps) => (
-    <StyledContent className={classnames('anchor-modal-content', className)}>
-        {children}
-    </StyledContent>
-);
-
-// Modal.Header
-// ------------------------------------------------------------------------------------------------------------------
-
-const StyledHeader = styled.div<ModalHeaderProps>`
-    box-sizing: border-box;
-    min-height: 4rem;
-    padding: 1.25rem 4.75rem 1.25rem 2rem;
-    width: 100%;
-    border-radius: ${borderRadius} ${borderRadius} 0 0;
-    margin-bottom: 1rem;
-
-    display: flex;
-    order: -1;
-
-    color: ${({ color = colors.charcoal.light }) => color};
-    background: ${({ background = colors.white.base }) => background};
-
-    // If the header exists, remove the Content's padding.
-    // We're preferring this selector to using Modal :first-child
-    // in case there is a close button placed before it, and using a
-    // double selector to ensure it takes precedence.
-    & ~ ${StyledContent}${StyledContent} {
-        padding-top: 0;
-    }
-`;
-
-interface ModalHeaderProps {
-    children?: any;
-    className?: string;
-    title?: string;
-    color?: string;
-    background?: string;
-}
-
-const ModalHeader = ({
-    children,
-    className,
-    title,
-    ...props
-}: ModalHeaderProps) => (
-    <StyledHeader
-        className={classnames('anchor-modal-header', className)}
-        {...props}
-    >
-        {title && (
-            <Typography scale={20} weight="bold">
-                {title}
-            </Typography>
-        )}
-        {children}
-    </StyledHeader>
-);
-
-// Modal.Footer
-// ------------------------------------------------------------------------------------------------------------------
-
-const StyledFooter = styled.div<ModalFooterProps>`
-    box-sizing: border-box;
-    width: 100%;
-    border-radius: 0 0 ${borderRadius} ${borderRadius};
-    margin-top: auto;
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    order: 1;
-    justify-self: flex-end;
-
-    background: ${({ background = 'transparent' }) => background};
-`;
-
-interface ModalFooterProps {
-    children: any;
-    className?: string;
-    background?: string;
-}
-
-const ModalFooter = ({ children, className, ...props }: ModalFooterProps) => (
-    <StyledFooter
-        className={classnames('anchor-modal-footer', className)}
-        {...props}
-    >
-        {children}
-    </StyledFooter>
-);
-
-// Modal.Close
-// ------------------------------------------------------------------------------------------------------------------
-
-const StyledClose = styled.div`
-    position: absolute;
-    height: 3rem;
-    width: 3rem;
-    margin: 0.5rem;
-    top: 0;
-    right: 0;
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    cursor: pointer;
-`;
-
-interface ModalCloseProps extends IconSVGProps {
-    onClick?: any;
-    className?: string;
-}
-
-const ModalClose = ({
-    onClick,
-    className,
-    color = colors.savvyCyan.base,
-    scale = 'lg',
-    ...props
-}: ModalCloseProps) => (
-    <StyledClose
-        className={classnames('anchor-modal-close', className)}
-        onClick={onClick}
-    >
-        {/* This may change to an icon-only button, once that component is done */}
-        <Close color={color} scale={scale} {...props} />
-    </StyledClose>
-);
-
-// Modal
-// ------------------------------------------------------------------------------------------------------------------
 
 const defaultSize: ModalSize = 'lg';
 
@@ -181,14 +32,14 @@ const StyledModal = StyledReactModal.default.styled`
         width || `${Sizes[size].width}rem`};
     height: ${({ height = '42.375rem' }: ModalProps) => height};
     margin: ${({ oversized, margin }: ModalProps) =>
-        margin || (oversized ? '2rem' : '0')};
+        margin || (oversized ? '2rem' : 'auto')};
 
     display: flex;
     align-items: center;
     flex-direction: column;
     justify-content: space-between;
 
-    border-radius: ${borderRadius};
+    border-radius: ${sizes.border.radius.modal};
     background: ${({ background = colors.white.base }: ModalProps) =>
         background};
     ${({ color }: ModalProps) =>
@@ -246,28 +97,23 @@ export const Modal = ({
     oversized,
     backgroundProps,
     ...props
-}: ModalProps): React.ReactElement<ModalProps> => {
-    return (
-        <StyledModal
-            {...{
-                className: classnames('anchor-modal', className),
-                oversized,
-                backgroundProps: { ...backgroundProps, oversized },
-                ...props,
-            }}
-        >
-            {children}
-        </StyledModal>
-    );
-};
+}: ModalProps): React.ReactElement<ModalProps> => (
+    <StyledModal
+        {...{
+            className: classnames('anchor-modal', className),
+            oversized,
+            backgroundProps: { ...backgroundProps, oversized },
+            ...props,
+        }}
+    >
+        {children}
+    </StyledModal>
+);
 
 Modal.Content = ModalContent;
 Modal.Header = ModalHeader;
 Modal.Footer = ModalFooter;
 Modal.Close = ModalClose;
-
-// BaseModalBackground
-// ------------------------------------------------------------------------------------------------------------------
 
 interface BaseModalBackgroundProps {
     opacity?: number;
@@ -279,8 +125,8 @@ const CustomModalBackground = styled(StyledReactModal.BaseModalBackground)`
         oversized &&
         css`
             align-items: initial;
-            overflow-y: scroll;
         `}
+    overflow-y: auto;
     background-color: ${({ opacity = 0.6 }: BaseModalBackgroundProps) =>
         opacify(opacity, 'rgba(0,0,0,0)')};
 `;
