+++ src/Modal/Modal.spec.tsx
@@ -1,12 +1,14 @@
 // REACT
 import * as React from 'react';
-import * as renderer from 'react-test-renderer';
+
 // ENZYME
 import { mount } from 'enzyme';
+
 // STYLED COMPONENTS
 /* tslint:disable no-import-side-effect*/
 import 'jest-styled-components';
 /* tslint:enable */
+
 // COMPONENT
 import { Modal, ModalProvider } from './Modal.component';
 const { Header, Content, Footer, Close } = Modal;
@@ -44,33 +46,5 @@ describe('Component: Modal', () => {
 
         const wrapper = mount(subject);
         expect(wrapper.html()).toMatchSnapshot();
-    });
-    describe('Component: Modal.Header', () => {
-        it('should render with title', () => {
-            const subject = <Header title="Header Title" />;
-            const tree = renderer.create(subject).toJSON();
-            expect(tree).toMatchSnapshot();
-        });
-    });
-    describe('Component: Modal.Content', () => {
-        it('should render', () => {
-            const subject = <Content>Text</Content>;
-            const tree = renderer.create(subject).toJSON();
-            expect(tree).toMatchSnapshot();
-        });
-    });
-    describe('Component: Modal.Footer', () => {
-        it('should render', () => {
-            const subject = <Footer>Text</Footer>;
-            const tree = renderer.create(subject).toJSON();
-            expect(tree).toMatchSnapshot();
-        });
-    });
-    describe('Component: Modal.Close', () => {
-        it('should render', () => {
-            const subject = <Close />;
-            const tree = renderer.create(subject).toJSON();
-            expect(tree).toMatchSnapshot();
-        });
     });
 });
